Version 1

+++ agenda/borrarActividad.php
@@ -1,22 +1,27 @@
 <?php
-
-
- include 'conexion.php';
-
-
- $id = $_GET["id"];
-
- $sql = "DELETE FROM actividades WHERE idAct='$id'";
-
- if($con -> query($sql)==TRUE){
-   
-
- }
-
-$con->close();
 include 'conexion.php';
 
+$id = $_GET["id"];
 
+// Verificar si el registro se está utilizando en la tabla agenda
+$sql_verificar = "SELECT COUNT(*) as cuenta FROM agenda WHERE idAct='$id'";
+$result_verificar = $con->query($sql_verificar);
+$cuenta = $result_verificar->fetch_assoc()["cuenta"];
+if ($cuenta > 0) {
+    echo "No se puede eliminar la actividad porque está siendo utilizada.";
+    exit();
+}
+
+$sql = "DELETE FROM actividades WHERE idAct='$id'";
+
+if($con -> query($sql)==TRUE){
+    // El registro se eliminó correctamente
+}
+
+$con->close();
+
+
+include 'conexion.php';
 
 $sql="SELECT * FROM actividades ORDER BY nombreActividad ASC ";        
 $resultset = $con->query($sql);
@@ -31,7 +36,4 @@ if($resultset->num_rows>0){
    </td></tr>"; 
  }
 }
-
 ?>
-
-
