+++ potreros/eliminarPotreros.php
@@ -1,23 +1,29 @@
 <?php
+include 'conexion.php';
 
+$id = $_GET["id"];
 
- include 'conexion.php';
+// Verificar si el registro se está utilizando en la tabla ganado
+$sql_verificar = "SELECT COUNT(*) as cuenta FROM ganado WHERE idPotrero='$id'";
+$result_verificar = $con->query($sql_verificar);
+$cuenta = $result_verificar->fetch_assoc()["cuenta"];
+if ($cuenta > 0) {
+    echo "No se puede eliminar el potrero porque está siendo utilizado.";
+    exit();
+}
 
- $id = $_GET["id"];
+$sql = "DELETE FROM potrero WHERE idPotrero='$id'";
 
- $sql = "DELETE FROM potrero WHERE idPotrero='$id'";
-
- if($con -> query($sql)==TRUE){
-   
-
- }
+if($con -> query($sql)==TRUE){
+    // El registro se eliminó correctamente
+}
 
 $con->close();
 
 
 include 'conexion.php';
 
-$sql="SELECT * FROM potrero ORDER BY nombrePotrero ASC";        
+$sql="SELECT * FROM potrero ORDER BY nombrePotrero ASC LIMIT 5";        
 $resultset = $con->query($sql);
 if($resultset->num_rows>0){
  while($fila = $resultset->fetch_assoc()){
@@ -34,7 +40,4 @@ if($resultset->num_rows>0){
     </td></tr>";  
 }
 }
-
 ?>
-
-
